requirements.txt
flask==0.10.1
mysql-python==1.2.5
